requirements.txt
coreapi==2.3.0
coreschema==0.0.4
Django==1.11
django-coreapi==0.5.0
django-cors-headers==2.0.2
django-extensions==1.7.9
django-filter==1.0.2
djangorestframework==3.6.2
djangorestframework-jwt==1.10.0
itypes==1.1.0
Jinja2==2.9.6
Markdown==2.6.8
MarkupSafe==1.0
psycopg2==2.7.1
Pygments==2.2.0
PyJWT==1.5.0
pytz==2017.2
requests==2.13.0
six==1.10.0
uritemplate==3.0.0
